Rename AuthWindow to AuthForm because it does not create a window.
